Renamed primary FlowCoordinator
